Move TagLineEdit from project_and_tags.py to tags.py
